further refactoring; made metadata cache globally accessible
